+++ src/main/java/org/cric/back_office/user/repository/RefreshTokenRepository.java
@@ -13,7 +13,7 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
 
     Optional<RefreshToken> findByUserId(Long userId);
 
-    void deleteByUserId(Integer userId);
+    void deleteByUserId(Long userId);
 
     void deleteByToken(String token);
 }
